+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### 0.2.0
+
+- Add sleep route (ericof)
+- Add error route (ericof)
+
 ### 0.1.0
 
 - Initial release (ericof)